ecom-bot/pyproject.toml
[project]
name = "ecom-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
]
